# server/src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API Computación_UTM",
        "description": "<html><body>Esta es la documentación de la API para un sistema de control de los Alumnos, Profesores y Salones de la Carrera de Computación en la UTM. <p>Para poder usar esta API es necesario solicitar el token y con esto poder utilizar los  servicios proporcionados.</p> <br> <p> <b>Autenticación y token de API</b></p><p>El administrador del sistema de la Carrera de Computación de la UTM proporciona el api-token una vez que haya creado y confirmado su cuenta.</p></body></html>",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:5000",
    "basePath": "",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/Computacion/Alumnos/{matricula}": {
            "get": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Muestra la información de un Alumno de la Carrera de Computación de la UTM dada su matrícula.",
                "description": "<p>Servicio que devuelve la información detallada de un los Alumno que pertenecen a la Carrera de Computación de la UTM </p><p>&nbsp;</p><p><em><strong>¿Cuáles son los Alumnos?</strong></em></p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "matricula": 1,
                                    "nombre": "Arturo Aquino",
                                    "edad": 21,
                                    "sexo": "masculino",
                                    "Grado": 5
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Alumnos"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "matricula",
                        "in": "path",
                        "required": true,
                        "description": "requiere matrícula del Alumno",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
        "/api/Computacion/Alumnos/update/{matricula}":{
            "put": {
                "summary": "Actualiza la información de un Alumno de la Carrera de Computación de la UTM dada su matrícula",
                "description": "<p>Servicio que actualiza la información del Alumno que pertenece a la Carrera de Computación de la UTM dada su matrícula.</p>",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "matricula",
                        "in": "path",
                        "description": "Ingrese la matrícula...",
                        "required": true,
                        "default": 1
                    },
                    {
                        "name": "Alumno",
                        "in": "body",
                        "description": "Ingrese la información del Alumno",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Alumnos"
                        }
                    }
                ],
                "tags": [
                    "Alumnos"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/Computacion/Alumnos/delete/{matricula}":{
            "delete": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Elimina la información de un Alumno de la Carrera de Computación de la UTM dada su matrícula.",
                "description": "<p>Servicio que elimina la información del Alumno que pertenecen a la Carrera de Computación de la UTM dada su matrícula.</p>",
                "operationId": "eliminar",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "matricula",
                        "in": "path",
                        "required": true,
                        "description": "requiere matrícula del Alumno",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
...
        "/api/Computacion/Profesores/{id}": {
            "get": {
                "tags": [
                    "Profesores"
                ],
                "summary": "Muestra la información de un Profesor de la Carrera de Computación de la UTM dado su id.",
                "description": "<p>Servicio que devuelve la información detallada de un Profesor perteneciente a la Carrera de Computación de la UTM </p><p>&nbsp;</p><p><em><strong>¿Cuál es el Profesor?</strong></em></p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "id": 1,
                                    "nombre": "Eduardo Soto",
                                    "edad": 50,
                                    "sexo": "masculino",
                                    "materias_impartidas": 2,
                                    "correo": "eduardosoto@gs.utm.mx",
                                    "tipo":2
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Profesores"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere id del Profesor",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
        "/api/Computacion/Profesores/delete/{id}":{
            "delete": {
                "tags": [
                    "Profesores"
                ],
                "summary": "Elimina la información de un Profesor de la Carrera de Computación de la UTM dado su id.",
                "description": "<p>Servicio que elimina la información del Profesor que pertenecen a la Carrera de Computación de la UTM dado su id.</p>",
                "operationId": "eliminar",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere id del Profesor",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
        "/api/Computacion/Profesores/update/{id}":{
            "put": {
                "summary": "Actualiza la información de un Profesor de la Carrera de Computación de la UTM dado su id",
                "description": "<p>Servicio que actualiza la información del Profesor que pertenece a la Carrera de Computación de la UTM dado su id.</p>",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ingrese el id...",
                        "required": true,
                        "default": 1
                    },
                    {
                        "name": "Profesor",
                        "in": "body",
                        "description": "Ingrese la información del Profesor",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Profesores"
                        }
                    }
                ],
                "tags": [
                    "Profesores"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/Computacion/Profesores/Horarios/{id}": {
            "get": {
                "tags": [
                    "Horarios Profesores"
                ],
                "summary": "Muestra la información de los Horarios de un Profesor de la Carrera de Computación de la UTM dado su id.",
                "description": "<p>Servicio que devuelve los Horarios detallados de un Profesor perteneciente a la Carrera de Computación de la UTM </p><p>&nbsp;</p><p><em><strong>¿Cuál Horario de Profesor se agrega?</strong></em></p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "id": 6,
                                    "nombre": "Moises Ramírez",
                                    "materias impartidas": "Computo Reconfigurable",
                                    "grado": 4,
                                    "num_salon": "Aula 7",
                                    "hora":"9:00am-10:00am"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Horarios_Profesores"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere id del Profesor",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
        "/api/Computacion/Profesores/Horario/delete/{id}":{
            "delete": {
                "tags": [
                    "Horarios Profesores"
                ],
                "summary": "Elimina el Horario de un Profesor de la Carrera de Computación de la UTM dado su id.",
                "description": "<p>Servicio que elimina el Horario de un Profesor que pertenecen a la Carrera de Computación de la UTM dado su id.</p>",
                "operationId": "eliminar",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere id del Profesor",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
        "/api/Computacion/Profesores/Horarios/update/{id}":{
            "put": {
                "summary": "Actualiza el Horario de un Profesor de la Carrera de Computación de la UTM dado su id",
                "description": "<p>Servicio que actualiza el Horario de un Profesor que pertenece a la Carrera de Computación de la UTM dado su id.</p>",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ingrese el id...",
                        "required": true,
                        "default": 1
                    },
                    {
                        "name": "Horario",
                        "in": "body",
                        "description": "Ingrese el Horario del Profesor",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Horarios_Profesores"
                        }
                    }
                ],
                "tags": [
                    "Horarios Profesores"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/Computacion/Alumnos/Horarios/{matricula}": {
            "get": {
                "tags": [
                    "Horarios Alumnos"
                ],
                "summary": "Muestra la información de los Horarios de un Alumno de la Carrera de Computación de la UTM dado su matrícula.",
                "description": "<p>Servicio que devuelve los Horarios detallados de un Alumno perteneciente a la Carrera de Computación de la UTM </p><p>&nbsp;</p><p><em><strong>¿Cuál Horario del Alumno se agrega?</strong></em></p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "matricula": 1,
                                    "nombre": "Arturo Aquino",
                                    "materia": "Computo Reconfigurable",
                                    "hora":"9:00am-10:00am",
                                    "salon":"Aula 8"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Horarios_Alumnos"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere matrícula del Alumno",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
        "/api/Computacion/Alumnos/Horarios/delete/{matricula}":{
            "delete": {
                "tags": [
                    "Horarios Alumnos"
                ],
                "summary": "Elimina el Horario de un Alumno de la Carrera de Computación de la UTM dada su matrícula.",
                "description": "<p>Servicio que elimina el Horario de un Alumno que pertenecen a la Carrera de Computación de la UTM dada su matrícula.</p>",
                "operationId": "eliminar",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere matrícula del Alumno",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
        "/api/Computacion/Alumnos/Horarios/update/{matricula}":{
            "put": {
                "summary": "Actualiza el Horario de un Alumno de la Carrera de Computación de la UTM dada su matrícula",
                "description": "<p>Servicio que actualiza el Horario de un Alumno que pertenece a la Carrera de Computación de la UTM dada su matrícula.</p>",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ingrese la matrícula...",
                        "required": true,
                        "default": 1
                    },
                    {
                        "name": "Horario",
                        "in": "body",
                        "description": "Ingrese el Horario del Alumno",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Horarios_Alumnos"
                        }
                    }
                ],
                "tags": [
                    "Horarios Alumnos"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/Computacion/Profesores/Tutorados/{id}": {
            "get": {
                "tags": [
                    "Tutorados"
                ],
                "summary": "Muestra la información de los Tutorados de la Carrera de Computación de la UTM dada su matrícula.",
                "description": "<p>Servicio que devuelve la información detallada de un Tutorado perteneciente a la Carrera de Computación de la UTM </p><p>&nbsp;</p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "tutor": "Moises Ramírez",
                                    "id_alu": 1,
                                    "nombre_alumno": "Arturo Aquino"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Tutorados"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere matrícula del Alumno",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
        "/api/Computacion/Profesores/Tutorados/delete/{id}":{
            "delete": {
                "tags": [
                    "Tutorados"
                ],
                "summary": "Elimina la información de un Tutorado de la Carrera de Computación de la UTM dada su matrícula.",
                "description": "<p>Servicio que elimina la información de un Tutorado que pertenece a la Carrera de Computación de la UTM dada su matrícula.</p>",
                "operationId": "eliminar",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere matrícula del Alumno",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
        "/api/Computacion/Profesores/Tutorados/update/{id}":{
            "put": {
                "summary": "Actualiza la información de un Tutorado de la Carrera de Computación de la UTM dada su matrícula",
                "description": "<p>Servicio que actualiza la información de un Tutorado que pertenece a la Carrera de Computación de la UTM dada su matrícula.</p>",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ingrese la matrícula...",
                        "required": true,
                        "default": 1
                    },
                    {
                        "name": "Tutoria",
                        "in": "body",
                        "description": "Ingrese la información de la Tutoría",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tutorados"
                        }
                    }
                ],
                "tags": [
                    "Tutorados"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/Computacion/Tutorias/delete/{id_al}":{
            "delete": {
                "tags": [
                    "Tutoria"
                ],
                "summary": "Elimina la información de una Tutoría de un Alumno de la Carrera de Computación de la UTM dada su matrícula.",
                "description": "<p>Servicio que elimina la información de una Tutoría a un Alumno que pertenece a la Carrera de Computación de la UTM dada su matrícula.</p>",
                "operationId": "eliminar",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere matrícula del Alumno",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
        "/api/Computacion/Tutorias/{id_al}": {
            "get": {
                "tags": [
                    "Tutoria"
                ],
                "summary": "Muestra la información de una Tutoría a un Alumno de la Carrera de Computación de la UTM dada su matrícula.",
                "description": "<p>Servicio que devuelve la información detallada de una Tutoría a un Alumno perteneciente a la Carrera de Computación de la UTM </p><p>&nbsp;</p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "tutor": "Moises Ramírez",
                                    "id_alu": 1,
                                    "nombre_alumno": "Arturo Aquino"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Tutoria"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere matrícula del Alumno",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
...
        "/api/Computacion/Tutorias/update/{id_al}":{
            "put": {
                "summary": "Actualiza la información de una Tutoría de la Carrera de Computación de la UTM dado su id",
                "description": "<p>Servicio que actualiza la información de una Tutoría dada a un Alumno de la Carrera de Computación de la UTM dado su id.</p>",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ingrese el id...",
                        "required": true,
                        "default": 1
                    },
                    {
                        "name": "Tutor",
                        "in": "body",
                        "description": "Ingrese la información del Tutor",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tutoria"
                        }
                    }
                ],
                "tags": [
                    "Tutoria"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/Computacion/Alumnos/Tutores/{id}": {
            "get": {
                "tags": [
                    "Tutores"
                ],
                "summary": "Muestra la información de los Tutores de la Carrera de Computación de la UTM dado su id.",
                "description": "<p>Servicio que devuelve la información detallada de un Tutor perteneciente a la Carrera de Computación de la UTM </p><p>&nbsp;</p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "id_prof": 1,
                                    "nombre_profesor": "Erik Ramos",
                                    "horario_asesoria": "5:00pm-6:00pm",
                                    "atutorados": "Omar Cruz"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Tutores"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere id del Tutor",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
        "/api/Computacion/Alumnos/Tutores/delete/{id}":{
            "delete": {
                "tags": [
                    "Tutores"
                ],
                "summary": "Elimina la información de un Tutor de la Carrera de Computación de la UTM dado su id.",
                "description": "<p>Servicio que elimina la información de un Tutor que pertenece a la Carrera de Computación de la UTM dado su id.</p>",
                "operationId": "eliminar",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere id del Tutor",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
        "/api/Computacion/Alumnos/Tutores/update/{id}":{
            "put": {
                "summary": "Actualiza la información de un Tutor de la Carrera de Computación de la UTM dado su id",
                "description": "<p>Servicio que actualiza la información de un Tutor que pertenece a la Carrera de Computación de la UTM dado su id.</p>",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ingrese el id...",
                        "required": true,
                        "default": 1
                    },
                    {
                        "name": "Tutor",
                        "in": "body",
                        "description": "Ingrese la información del Tutor",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tutores"
                        }
                    }
                ],
                "tags": [
                    "Tutores"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Alumnos": {
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string",
                    "description": "Describe el nombre de los Alumnos"
                },
                "edad": {
                    "type": "number",
                    "description": "Describe la edad de los Alumnos"
                },
                "sexo": {
                    "type": "string",
                    "description": "Describe el sexo de los Alumnos"
                },
                "grado": {
                    "type": "number",
                    "description": "Describe el grado de los Alumnos"
                },
                "grupo": {
                    "type": "string",
                    "description": "Describe el grupo de los Alumnos"
                },
                "materias":{
                    "type": "string",
                    "description": "Describe las materias que lleva el Alumno"
                },
                "puntaje":{
                    "type": "string",
                    "description": "Describe las calificaciones del Alumno"
                },
                "promedio":{
                    "type": "number",
                    "description": "Describe el promedio del Alumno"
                }
            }
        },
...
        "Tutorados": {
            "type": "object",
            "properties": {
                "tutor":{
                    "type": "string",
                    "description": "Describe el nombre del Tutor del Tutorado"
                },
                "id_alu": {
                    "type": "number",
                    "description": "Identificador único de un Tutorado de la Carrera de Computación de la UTM"
                },
                "nombre_alumno":{
                    "type": "string",
                    "description": "Describe el nombre del Tutorado"
                }
            }
        },
...
    }
}